pyproject.toml
[tool.poetry]
name = "dmlutgard"
version = "0.1.0"
description = ""
authors = ["yobmod <yobmod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "2.2.24"
dj-database-url = "*"
django-dotenv = "*"
django-storages = {extras = ["boto3"], version = "*"}
psycopg2 = "*"
CherryPy = "*"
gunicorn = "*"
whitenoise = "*"
django-extensions = "*3"
djangorestframework = "*"
django-compressor = "*"
django-libsass = "^0.9"

[tool.poetry.dev-dependencies]
Werkzeug = "*"  # -> for extensions
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
pytest-sugar = "*"
pytest-xdist = "*"
mypy = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
autopep8 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
